Added nginx config and configured flask jwt extended

+++ resources/item.py
@@ -1,34 +1,44 @@
 import sqlite3
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import (
+    jwt_required,
+    get_jwt_claims,
+    jwt_optional,
+    get_jwt_identity,
+    fresh_jwt_required,
+)
 from models.item import ItemModel
+
+BLANK_ERROR = "This {} cannot be left blank!"
+ITEM_NOT_FOUND = "Item: {} not found"
+NAME_ALREADY_EXISTS = "An item with name '{}' already exists."
+NOT_FOUND = "{} doesn't exist!"
+ITEM_DELETED =  "{} deleted"
 
 class Item(Resource):
     parser = reqparse.RequestParser()
-    parser.add_argument('price',
-        type = float,
-        required = True,
-        help = "This field cannot be left blank."
+    parser.add_argument(
+        "price", type=float, required=True, help=BLANK_ERROR.format("price")
     )
-    parser.add_argument('store_id',
-        type = int,
-        required = True,
-        help = "Every item need a store id."
+    parser.add_argument(
+        "store_id", type=int, required=True, help=BLANK_ERROR.format("store_id")
     )
 
-    @jwt_required()
-    def get(self, name):
+    @classmethod
+    @jwt_required
+    def get(cls, name: str):
         item = ItemModel.find_by_name(name)
         if item:
             return item.json()
-        return {'message': f"Item: '{name}' not found"}, 404
+        return {"message": ITEM_NOT_FOUND.format(name)}, 404
 
-
-    def post(self, name):
+    @classmethod
+    @fresh_jwt_required
+    def post(cls, name: str):
         if ItemModel.find_by_name(name):
-            return {'message': f"An item with name '{name}' already exists."}, 400
+            return {"message": NAME_ALREADY_EXISTS.format(name)}, 400
         data = Item.parser.parse_args()
-        
+
         item = ItemModel(name, **data)
 
         try:
@@ -38,30 +48,54 @@ class Item(Resource):
 
         return item.json(), 201
 
+    @classmethod
+    @jwt_required
+    def delete(cls, name: str):
+        claims = get_jwt_claims()
+        if not claims["is_admin"]:
+            return {"message": "Admin privilege required!"}, 401
 
-    def delete(self, name):
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
-            return {'message': f"{name} deleted"}
-        return {'message': f"{name} doesn't exist!"}
+            return {"message": ITEM_DELETED.format(name)}
+        return {"message": NOT_FOUND.format(name)}
     
-
-    def put(self, name):
+    @classmethod
+    def put(cls, name: str):
         data = Item.parser.parse_args()
         item = ItemModel.find_by_name(name)
 
         if item is None:
             item = ItemModel(name, **data)
         else:
-            item.price = data['price']
+            item.price = data["price"]
 
         item.save_to_db()
-        
+
         return item.json()
-  
+
+
 class ItemList(Resource):
-    def get(self):
-        #items = [item.json() for item in ItemModel.find_all()]
-        items = list(map(lambda x: x.json(), ItemModel.find_all()))
-        return {'items' : items}
+
+    """
+    jwt_optional can be use to capture user_id 
+    and also show different data for authorized
+    and unauthorized users.
+    """
+
+    @classmethod
+    @jwt_optional
+    def get(cls):
+        user_id = get_jwt_identity()
+        items = [item.json() for item in ItemModel.find_all()]
+        # items = list(map(lambda x: x.json(), ItemModel.find_all()))
+        if user_id:
+            return {"items": items}, 200
+        return (
+            {
+                "items": [item["name"] for item in items],
+                "message": "Please to login to view item detail",
+            },
+            200,
+        )
